Backend/requirements.txt
patool
dataclasses==0.6
Django==3.1.4
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pytest-asyncio==0.14.0
pytest-django==4.1.0
django-cors-headers
django-filter
requests
django-allauth
psycopg2
django-jsonfield
jsonfield
django-rest-auth
django-elasticsearch-dsl==7.1.4
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
celery
redis
sqlalchemy




